+++ fuzz/fuzz_targets/block_wire.rs
@@ -1,0 +1,8 @@
+#![no_main]
+
+use libfuzzer_sys::fuzz_target;
+use rbitcoin_consensus::check_block_wire;
+
+fuzz_target!(|data: &[u8]| {
+    let _ = check_block_wire(data);
+});